requirements.txt
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.3.2
filelock==3.0.12
gdown==3.12.2
h5py==2.10.0
idna==2.10
joblib==0.17.0
mkl-fft
mkl-random
numpy==1.19.2
olefile==0.46
pandas==1.1.3
Pillow==8.1.1
pip==20.2.4
pycparser==2.20
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
setuptools
six==1.15.0
threadpoolctl==2.1.0
#torch==1.4.0
#torchvision==0.5.0
tqdm==4.51.0
urllib3==1.25.11
wheel==0.35.1


argparse
easydict
matplotlib
open3d==0.9
opencv-python
pyyaml
tensorboardX
timm==0.4.5 
transforms3d
termcolor
